feat(common_contract): add CommonContract for handling common contract operations
refactor(incheon_contents): use CommonContract for IncheonContentsContract

fix(wallets): add conditional async_trait attribute for KaiaWallet implementations
