+++ packages/dto/src/contracts/mod.rs
@@ -1,2 +1,3 @@
+pub mod common_contract;
 pub mod incheon_contents;
 pub mod klaytn_transaction;
